requirements.txt
asgiref==3.8.1
tzdata==2024.1
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.7.0
python-dotenv==1.0.1
pytz==2024.1